Refactor code to make it more manageble
