+++ .gitignore
@@ -3,3 +3,6 @@
 ################################################################################
 
 /.vs/FullscreenLock/v15
+/.vs/FullscreenLock/v14
+/FullscreenLock/obj
+/FullscreenLock/bin
