+++ FullscreenLock/FullscreenLock.csproj
@@ -52,7 +52,8 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Checker.cs" />
+    <Compile Include="UIControllers\IMainForm.cs" />
+    <Compile Include="UIControllers\MainformController.cs" />
     <Compile Include="RECT.cs" />
     <Compile Include="FullscreenLock.cs">
       <SubType>Form</SubType>
@@ -91,6 +92,7 @@
   <ItemGroup>
     <Content Include="gaben.ico" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
