+++ FullscreenLock/UIControllers/IMainForm.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace FullscreenLock.UIControllers
+{
+    public interface IMainForm
+    {
+        void SetInformation(string info);
+    }
+}
